Ahora se envía de vuelta al cliente un mensaje si la conexión tiene éxito.

+++ cliente.cpp
@@ -65,7 +65,7 @@ int main(int argc, const char * argv[])
 		exit(0);
 	}
 
-	rc = recv(sock, buffer, sizeof(int), 0);
+	rc = recv(sock, buffer, sizeof(uint8_t), 0);
 
 	struct mensaje_saludo nuevo_saludo;
 	string s = "Jordi";
